moved powerup spawning code out of main
