+++ static/game/spawner.js
@@ -1,18 +1,20 @@
 import {
   addToWorld,
   cellToWorldPosition,
-  getImportantEntity
+  getImportantEntity,
+  getDimensions
 } from "../modules/gamemanager.js";
 import { randomInt, griderate } from "../modules/helpers.js";
 import { Chase } from "./chase.js";
 import { Crosser } from "./crosser.js";
-import { distanceBoard } from "./generator.js";
+import { distanceBoard, segregateTerrain } from "./generator.js";
 import { Scatter } from "./scatter.js";
 import { Shooter } from "./shooter.js";
 import { Bomber } from "./bomber.js";
 import { Vector } from "../modules/vector.js";
 import { ChanceTable } from "../modules/chancetable.js";
 import { Enemy } from "./enemy.js";
+import { powerUpTypes } from "./powerups/poweruptypes.js";
 
 /**
  * a continuous function that 0 when x is below a certain point, 1 when x is
@@ -93,4 +95,76 @@ export function spawnEnemies(
       console.log("added an enemy to the world");
     }
   });
+}
+
+/**
+ * spawn powerups into the world
+ * @param {number[][]} board 
+ * @param {number} [powerupChance] not 0 to 1, but rather in the hundreds
+ */
+export function spawnPowerups(board, powerupChance = 280) {
+  const {width: blockWidth, height: blockHeight} = getDimensions();
+
+  // Get the segregated board
+  const {
+    segregatedBoard: segregatedBoard,
+    groupNum: groupNum,
+    largestGroup: largestGroup
+  } = segregateTerrain(board);
+
+  // Init the cave locations array
+  const caveLocations = [];
+  for (let i = 0; i < groupNum; i++) {
+    caveLocations.push([]);
+  }
+
+  // For each cave, give it a list of available terrain.
+  for (let i = 0; i < segregatedBoard.length; i++) {
+    for (let j = 0; j < segregatedBoard[i].length; j++) {
+      const location = segregatedBoard[i][j];
+      if (location != 0)
+        caveLocations[location - 1].push(
+          new Vector(i * blockWidth, j * blockHeight)
+        );
+    }
+  }
+
+  // If empty numbers exist (They shouldn't, but do) delete them.
+  // TODO: figure out why we need this
+  for (let i = 0; i < caveLocations.length; i++) {
+    if (caveLocations[i].length == 0) caveLocations.splice(i, i);
+  }
+  const tilesPerAdditionalPowerupChance = powerupChance;
+
+  for (let i = 0; i < caveLocations.length; i++) {
+    if (i == largestGroup) continue;
+
+    // Have a chance for an additional powerup for every 10 blocks.
+    const additional_powerups = Math.floor(
+      Math.max(1000 - caveLocations[i].length, 0) /
+        tilesPerAdditionalPowerupChance
+    );
+    const powerup_num = Math.floor(Math.random() * additional_powerups) + 1;
+
+    for (let p = 0; p < powerup_num; p++) {
+      const randomIndex = Math.floor(Math.random() * caveLocations[i].length);
+      const randomTile = caveLocations[i][randomIndex];
+      caveLocations[i].splice(randomIndex, 1);
+
+      const location = randomTile.add(
+        new Vector(blockWidth / 2, blockHeight / 2)
+      );
+
+      const r = Math.random();
+      const magnitude = Math.floor(Math.random() * 5) + 1;
+      for (let j = 1; j <= powerUpTypes.length; ++j) {
+        if (r < j / powerUpTypes.length) {
+          const powerUp = new powerUpTypes[j - 1](magnitude, location);
+          addToWorld(powerUp);
+          break;
+        }
+      }
+    }
+  }
+
 }
